server/requirements.txt
fastapi==0.68
sqlalchemy==1.4
uvicorn==0.15
python-multipart
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
jinja2==3.0.1
aiofiles==0.7.0
requests==2.26.0
websockets==10.0